# api.yaml
openapi: 3.1.0
x-stoplight:
  id: kyv86omeutrhp
info:
  title: CalenGo
  version: '1.0'
servers: []
paths:
  '/api/v1/note/{note}':
    get:
      security:
        - BasicAuth: []
      summary: Your GET endpoint
      tags: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Note'
        '400':
          description: Bad Request
      operationId: get-api-v1-note
      x-stoplight:
        id: ho4yimsfhex80
      x-internal: false
      parameters:
        - name: note
          schema:
            type: string
          in: path
          required: true
  /api/v1/note:
    post:
      security:
        - BasicAuth: []
      summary: Your POST endpoint
      tags: []
      responses:
        '200':
          description: OK
        '400':
          description: Bad Request
      operationId: post-api-v1-note
      x-stoplight:
        id: scnt530gfp1wr
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Note'
      parameters: []
    put:
      security:
        - BasicAuth: []
      summary: Your POST endpoint
      tags: []
      responses:
        '200':
          description: OK
        '400':
          description: Bad Request
      operationId: put-api-v1-note
      x-stoplight:
        id: v55qh58ya26ch
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Note'
      parameters: []
    delete:
      security:
        - BasicAuth: []
      summary: Your POST endpoint
      tags: []
      responses:
        '200':
          description: OK
        '400':
          description: Bad Request
      operationId: delete-api-v1-note
      x-stoplight:
        id: x605v0i649cnf
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Note'
      parameters: []
    patch:
      security:
        - BasicAuth: []
      summary: Your GET endpoint
      tags: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Note'
                  x-stoplight:
                    id: fzvhu0pec30is
        '400':
          description: Bad Request
      operationId: list-api-v1-note
      x-stoplight:
        id: b6kw93bcad9b7
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required:
                - startTime
                - endTime
              properties:
                startTime:
                  type: string
                  x-stoplight:
                    id: hn7yledguual5
                endTime:
                  type: string
                  x-stoplight:
                    id: yqtljllphf9wx
      parameters: []
  /api/v1/conflict:
    get:
      security:
        - BasicAuth: []
      summary: Your GET endpoint
      tags: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  x-stoplight:
                    id: mp0vtbpllhpd7
                  type: object
                  required:
                    - first
                    - second
                  properties:
                    first:
                      $ref: '#/components/schemas/Note'
                      x-stoplight:
                        id: uupsf75x28wyu
                    second:
                      $ref: '#/components/schemas/Note'
                      x-stoplight:
                        id: f35zbaokatt1u
        '400':
          description: Bad Request
      operationId: conflict-api-v1
      x-stoplight:
        id: ho4yimsfhex85
      x-internal: false
  '/api/v1/user/{user}':
    get:
      security:
        - BasicAuth: []
      summary: Your GET endpoint
      tags: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Bad Request
      operationId: get-api-v1-user
      x-stoplight:
        id: jltqzwem9twl5
      x-internal: false
      parameters:
        - name: user
          schema:
            type: string
          in: path
          required: true
  /api/v1/user:
    post:
      summary: Your POST endpoint
      tags: []
      responses:
        '200':
          description: OK
        '400':
          description: Bad Request
      operationId: post-api-v1-user
      x-stoplight:
        id: 8wk48xcvo5zxw
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      parameters: []
    put:
      security:
        - BasicAuth: []
      summary: Your POST endpoint
      tags: []
      responses:
        '200':
          description: OK
        '400':
          description: Bad Request
      operationId: put-api-v1-user
      x-stoplight:
        id: dao0qiyhjj4r5
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      parameters: []
    delete:
      summary: Your POST endpoint
      tags: []
      responses:
        '200':
          description: OK
        '400':
          description: Bad Request
      operationId: delete-api-v1-user
      x-stoplight:
        id: 1niyz8w4gde1x
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      parameters: []
components:
  securitySchemes:
    BasicAuth:
      type: http
      scheme: basic
  schemas:
    User:
      title: User
      type: object
      x-stoplight:
        id: t15cxa63lqsmu
      required:
        - name
        - password
      properties:
        name:
          type: string
          x-stoplight:
            id: 4p2cygg8fj5gq
          x-oapi-codegen-extra-tags:
            db: name
        password:
          type: string
          x-stoplight:
            id: oj84mr4oak9ve
          x-oapi-codegen-extra-tags:
            db: password
    Note:
      title: Note
      x-stoplight:
        id: 4tmzl66n9aymk
      type: object
      required:
        - name
        - description
        - startTime
        - endTime
      properties:
        name:
          type: string
          x-stoplight:
            id: gang5v8gpmm2o
          x-oapi-codegen-extra-tags:
            db: name
        description:
          type: string
          x-stoplight:
            id: grdf06u2i8me1
          x-oapi-codegen-extra-tags:
            db: description
        startTime:
          type: string
          x-stoplight:
            id: r6na1bfsjea5l
          x-oapi-codegen-extra-tags:
            db: start_time
        endTime:
          type: string
          x-stoplight:
            id: fo4g30g8ggcd2
          x-oapi-codegen-extra-tags:
            db: end_time
  parameters: {}